+++ src/main/java/com/snowofsunflower/android/image/ImageRound.kt
@@ -4,5 +4,5 @@ package com.snowofsunflower.android.image
 /**
  * 圆角图形
  */
-data class ImageRound(val round: Int) : IImageShape
+data class ImageRound(val round: Float) : IImageShape
